+++ src/nationalarchives/components/global-header/global-header.stories.js
@@ -10,6 +10,7 @@ const argTypes = {
   collapseOnMedium: { control: "boolean" },
   navigationId: { control: "text" },
   topNavigationId: { control: "text" },
+  phaseBanner: { control: "object" },
   classes: { control: "text" },
   attributes: { control: "object" },
 };
@@ -35,6 +36,7 @@ const Template = ({
   collapseOnMedium,
   navigationId,
   topNavigationId,
+  phaseBanner,
   classes,
   attributes,
 }) =>
@@ -46,6 +48,7 @@ const Template = ({
       collapseOnMedium,
       navigationId,
       topNavigationId,
+      phaseBanner,
       classes,
       attributes,
     },
@@ -365,4 +368,59 @@ Mobile.play = async ({ args, canvasElement, step }) => {
   // });
 
   $navigationToggle.blur();
+};
+
+export const WithPhaseBanner = Template.bind({});
+WithPhaseBanner.args = {
+  logo: {
+    href: "#/",
+  },
+  topNavigation: [
+    {
+      text: "Search",
+      href: "https://www.nationalarchives.gov.uk/search/",
+      icon: "search",
+    },
+    {
+      text: "Shop",
+      href: "https://shop.nationalarchives.gov.uk/",
+      icon: "shop",
+    },
+    {
+      text: "Sign in",
+      href: "#/sign-in",
+      icon: "user",
+    },
+  ],
+  navigation: [
+    {
+      text: "Visit",
+      href: "https://www.nationalarchives.gov.uk/about/visit-us/",
+    },
+    {
+      text: "What’s on",
+      href: "https://www.nationalarchives.gov.uk/about/visit-us/whats-on/",
+    },
+    {
+      text: "Explore the collection",
+      href: "https://beta.nationalarchives.gov.uk/explore-the-collection/",
+    },
+    {
+      text: "Help using the archive",
+      href: "https://www.nationalarchives.gov.uk/help-with-your-research/",
+    },
+    {
+      text: "Education",
+      href: "https://www.nationalarchives.gov.uk/education/",
+    },
+    {
+      text: "Professional guidance and services",
+      href: "https://www.nationalarchives.gov.uk/professional-guidance-and-services/",
+    },
+  ],
+  phaseBanner: {
+    phase: "beta",
+    message: `This is a new service – <a href="#">give us your feedback</a> to help improve it.`,
+  },
+  classes: "tna-global-header--demo",
 };